+++ src/main/java/com/drone/drone/entity/Drone.java
@@ -34,7 +34,7 @@ public class Drone {
     @Column(nullable = false) //Add max and min limit
     @Max(500)
     @Min(0)
-    private double weightLimit;
+    private int weightLimit;
 
     @Column(nullable = false) //Add max and min limit
     @Max(100)
@@ -45,7 +45,7 @@ public class Drone {
     @Column(nullable = false)
     private DroneState state;
 
-    public Drone(@Size(max = 100) String serialNumber, DroneModel model, @Max(500) @Min(0) double weightLimit, @Max(100) @Min(0) int batteryCapacity, DroneState state) {
+    public Drone(@Size(max = 100) String serialNumber, DroneModel model, @Max(500) @Min(0) int weightLimit, @Max(100) @Min(0) int batteryCapacity, DroneState state) {
         this.serialNumber = serialNumber;
         this.model = model;
         this.weightLimit = weightLimit;
